Added RuntimeData test cases for: getJob() updateJob() delJob() get_attribute() set_attribute()
Also made small tweaks to some of said functions to more accurately
check the return values of the functions they call. For example:
instead of checking the bool() interpretation of the value returned
by a function, we explicitly check its value for None.

Also we make further use of _is_job_dict() in updateJob().